# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/problem-bundles:
    get:
      operationId: listProblemBundles
      tags: [admin]
      summary: List all problem bundles
      responses:
        "200":
          description: List of problem bundles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemBundle"
    post:
      operationId: createProblemBundle
      tags: [admin]
      summary: Create a problem bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemBundleCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemBundle"
  /admin/problem-bundles/{id}:
    patch:
      operationId: updateProblemBundle
      tags: [admin]
      summary: Update a problem bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemBundleCreate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemBundle"
    delete:
      operationId: deleteProblemBundle
      tags: [admin]
      summary: Delete a problem bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/problem-bundles/{id}/problems:
    get:
      operationId: listBundleProblems
      tags: [admin]
      summary: List problems in a bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problems in bundle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BundleProblemItem"
    post:
      operationId: addBundleProblem
      tags: [admin]
      summary: Add a problem to a bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBundleProblemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                type: object
  /admin/problem-bundles/{id}/problems/{problemId}:
    delete:
      operationId: removeBundleProblem
      tags: [admin]
      summary: Remove a problem from a bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: problemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /stats/video-duration:
    get:
      operationId: getVideoDuration
      tags: [stats]
      summary: Get video watch duration by day (weekly)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [weekly, monthly]
      responses:
        "200":
          description: Video duration data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoDurationEntry"
  /stats/module-completion:
    get:
      operationId: getModuleCompletion
      tags: [stats]
      summary: Get module completion calendar for current month
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Module completion data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModuleCompletionEntry"
...
components:
  schemas:
...
    VideoDurationEntry:
      type: object
      required: [date, minutes]
      properties:
        date:
          type: string
        minutes:
          type: integer
    ModuleCompletionEntry:
      type: object
      required: [day, completed]
      properties:
        day:
          type: integer
        completed:
          type: boolean
...
    ProblemBundle:
      type: object
      required: [id, title, createdAt, problemCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        problemCount:
          type: integer
    ProblemBundleCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
    BundleProblemItem:
      type: object
      required: [id, bundleId, problemId, orderIndex, title]
      properties:
        id:
          type: integer
        bundleId:
          type: integer
        problemId:
          type: integer
        orderIndex:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        totalMarks:
          type: integer
    AddBundleProblemInput:
      type: object
      required: [problemId]
      properties:
        problemId:
          type: integer
        orderIndex:
          type: integer
...